refactor: move out deployment-loader to own type file
