+++ packages/core/src/new-api/types/deployment-loader.ts
@@ -1,0 +1,19 @@
+import { Artifact, BuildInfo } from "./artifact";
+import { Journal } from "./journal";
+
+/**
+ * Read and write to the deployment storage.
+ *
+ * @beta
+ */
+export interface DeploymentLoader {
+  journal: Journal;
+  initialize(deploymentId: string): Promise<void>;
+  loadArtifact(storedArtifactPath: string): Promise<Artifact>;
+  storeArtifact(futureId: string, artifact: Artifact): Promise<string>;
+  storeBuildInfo(buildInfo: BuildInfo): Promise<string>;
+  recordDeployedAddress(
+    futureId: string,
+    contractAddress: string
+  ): Promise<void>;
+}
